feat: added service for geting one item from backend, based on the item id.
