+++ frontend/db.json
@@ -10,5 +10,67 @@
     { "id": 8, "title": "Item 3", "imageUrl": "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcT2Ld2-jqasWZCNTmx0oWiYdrSz8XSpUzAeQQ&s", "price": 150 },
     { "id": 9, "title": "Item 4", "imageUrl": "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSmmKsSJQM7-pvAYClePZmpp6oLWKHCn4rujA&s", "price": 250 },
     { "id": 10, "title": "Item 5", "imageUrl": "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSmmKsSJQM7-pvAYClePZmpp6oLWKHCn4rujA&s", "price": 300 }
+  ],
+  "item-detail": [
+    {
+      "id": 1,
+      "title": "Item 1",
+      "description": "This is a detailed description of Item 1.",
+      "price": 100,
+      "location": "New York, USA",
+      "contact": "+1 234 567 890",
+      "imageUrls": [
+        "https://upload.wikimedia.org/wikipedia/commons/f/f9/Phoenicopterus_ruber_in_S%C3%A3o_Paulo_Zoo.jpg",
+        "https://upload.wikimedia.org/wikipedia/commons/f/f9/Phoenicopterus_ruber_in_S%C3%A3o_Paulo_Zoo.jpg"
+      ]
+    },
+    {
+      "id": 2,
+      "title": "Item 2",
+      "description": "This is a detailed description of Item 2.",
+      "price": 200,
+      "location": "Los Angeles, USA",
+      "contact": "+1 987 654 321",
+      "imageUrls": [
+        "https://upload.wikimedia.org/wikipedia/commons/3/3a/Cat03.jpg",
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSmmKsSJQM7-pvAYClePZmpp6oLWKHCn4rujA&s"
+      ]
+    },
+    {
+      "id": 3,
+      "title": "Item 3",
+      "description": "This is a detailed description of Item 3.",
+      "price": 150,
+      "location": "London, UK",
+      "contact": "+44 20 7946 0958",
+      "imageUrls": [
+        "https://upload.wikimedia.org/wikipedia/commons/3/3a/Cat03.jpg",
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSmmKsSJQM7-pvAYClePZmpp6oLWKHCn4rujA&s"
+      ]
+    },
+    {
+      "id": 4,
+      "title": "Item 4",
+      "description": "This is a detailed description of Item 4.",
+      "price": 180,
+      "location": "Tokyo, Japan",
+      "contact": "+81 3-1234-5678",
+      "imageUrls": [
+        "https://upload.wikimedia.org/wikipedia/commons/3/3a/Cat03.jpg",
+        "https://upload.wikimedia.org/wikipedia/commons/3/3a/Cat03.jpg"
+      ]
+    },
+    {
+      "id": 5,
+      "title": "Item 5",
+      "description": "This is a detailed description of Item 5.",
+      "price": 220,
+      "location": "Sydney, Australia",
+      "contact": "+61 2 1234 5678",
+      "imageUrls": [
+        "https://upload.wikimedia.org/wikipedia/commons/3/3a/Cat03.jpg",
+        "https://upload.wikimedia.org/wikipedia/commons/3/3a/Cat03.jpg"
+      ]
+    }
   ]
 }
